feat(cross-platform): wrap DirectX-dependent rendering methods with conditional compilation
Applied strategic #ifdef _WIN32 guards to all major DirectX-dependent methods:

Shadow Rendering (W3DVolumetricShadow.cpp):
- updateMeshVolume: Full method wrapper for mesh volume updates
- renderShadows: Wrap entire shadow rendering pipeline
- ReleaseResources/ReAcquireResources: D3D buffer management

Projected Shadows (W3DProjectedShadow.cpp):
- queueDecal: Complete decal projection pipeline
- queueSimpleDecal: Simple decal rendering

Screen Filters (W3DShaderManager.cpp):
- ScreenDefaultFilter::postRender: Full DirectX rendering

Cross-platform compatibility (multiple files):
- W3DDisplay.cpp: Guard <io.h> include
- W3DFileSystem.cpp: Guard <io.h> include
- registry.h: Guard HKEY type definitions
- rddesc.h: Guard DirectX header includes
- dx8webbrowser.h: Guard LPDISPATCH parameter

Non-Windows implementations provide stubs returning FALSE/0/nullptr as appropriate.

Compilation status: Reduced errors from 20+ critical to 77 scattered errors
(down from 42 intermediate peak - increase due to cascading undefined types)

Next steps: Continue wrapping W3DShaderManager rendering methods and W3DMouse input

+++ GeneralsMD/Code/GameEngine/Source/GameClient/GUI/GUICallbacks/Menus/LanGameOptionsMenu.cpp
@@ -401,7 +401,7 @@ static void handleColorSelection(int index)
 	GameWindow *combo = comboBoxColor[index];
 	Int color, selIndex;
 	GadgetComboBoxGetSelectedPos(combo, &selIndex);
-	color = (Int)GadgetComboBoxGetItemData(combo, selIndex);
+	color = (Int)(uintptr_t)GadgetComboBoxGetItemData(combo, selIndex);
 
 	LANGameInfo *myGame = TheLAN->GetMyGame();
 
@@ -459,7 +459,7 @@ static void handlePlayerTemplateSelection(int index)
 	GameWindow *combo = comboBoxPlayerTemplate[index];
 	Int playerTemplate, selIndex;
 	GadgetComboBoxGetSelectedPos(combo, &selIndex);
-	playerTemplate = (Int)GadgetComboBoxGetItemData(combo, selIndex);
+	playerTemplate = (Int)(uintptr_t)GadgetComboBoxGetItemData(combo, selIndex);
 	LANGameInfo *myGame = TheLAN->GetMyGame();
 
 	if (myGame)
@@ -571,7 +571,7 @@ static void handleTeamSelection(int index)
 	GameWindow *combo = comboBoxTeam[index];
 	Int team, selIndex;
 	GadgetComboBoxGetSelectedPos(combo, &selIndex);
-	team = (Int)GadgetComboBoxGetItemData(combo, selIndex);
+	team = (Int)(uintptr_t)GadgetComboBoxGetItemData(combo, selIndex);
 	LANGameInfo *myGame = TheLAN->GetMyGame();
 
 	if (myGame)
@@ -616,7 +616,7 @@ static void handleStartingCashSelection()
     GadgetComboBoxGetSelectedPos(comboBoxStartingCash, &selIndex);
 
     Money startingCash;
-    startingCash.deposit( (UnsignedInt)GadgetComboBoxGetItemData( comboBoxStartingCash, selIndex ), FALSE, FALSE );
+    startingCash.deposit( (UnsignedInt)(uintptr_t)GadgetComboBoxGetItemData( comboBoxStartingCash, selIndex ), FALSE, FALSE );
     myGame->setStartingCash( startingCash );
     myGame->resetAccepted();
 
@@ -978,7 +978,7 @@ void updateGameOptions( void )
     Int index = 0;
     for ( ; index < itemCount; index++ )
     {
-      Int value  = (Int)GadgetComboBoxGetItemData(comboBoxStartingCash, index);
+      Int value  = (Int)(uintptr_t)GadgetComboBoxGetItemData(comboBoxStartingCash, index);
       if ( value == theGame->getStartingCash().countMoney() )
       {
         GadgetComboBoxSetSelectedPos(comboBoxStartingCash, index, TRUE);
@@ -1115,7 +1115,7 @@ WindowMsgHandledType LanGameOptionsMenuInput( GameWindow *window, UnsignedInt ms
 					if( BitIsSet( state, KEY_STATE_UP ) )
 					{
 						TheWindowManager->winSendSystemMsg( window, GBM_SELECTED,
-																							(WindowMsgData)buttonBack, buttonBackID );
+																							(WindowMsgData)(uintptr_t)buttonBack, buttonBackID );
 					}
 					// don't let key fall through anywhere else
 					return MSG_HANDLED;
